Sorting attempt
Sorry for all of the extra comments. Some changes:

A) Added a field in node that shows the lowercase version of the word (so we don't have to keep converting at string comparison)
B) Able to successfully extract words from buffer into nodes

My only problem is insert. It starts to work until I hit a case where the word is in the middle, and then I think the pointers get a little messed up. I think I might try to fix this by adjusting the for loop and conditionals (check for all three conditions: insert at head, insert in middle, insert at tail).

I also added another struct called linked_list, though I don't use it here. I may choose to use it, though, so as not to easily confuse inserted and comparison nodes with the one representing the whole list.